Extended JS behaviors to fix field cardinality handling.
When a media reference field limits the cardinality the media browser is
not adapting the JS validation behavior to limit the allowed selected
items to the constaints configured at the field instance. The
file_entity_browser module had the same issue where so we ported the
patch from https://www.drupal.org/project/file_browser/issues/2917823 to
work on media entity browser and adapted the file counts message to
display in single references source plugin view and added a limit
message warning when user reach the maximum selected items.